Address comments and code refactoring
